scripts(cf-push): display app logs on push error
As it is not possible to know which app is deployed, we have to parse
cf push logs. We extracts the log command from logs.

we enable pipefail. According to [1]: I want to add that when pipefail is on, it is enough for shell to think that the entire pipe has non-zero exit code when one of commands in the pipe has non-zero exit code (with pipefail off it must the last one).

 ```
 $ set -o pipefail
 $ false | true ; echo $?
 1
 $ set +o pipefail
 $ false | true ; echo $?
 0
 ```

[1] https://stackoverflow.com/questions/11231937/bash-ignoring-error-for-a-particular-command

close #5