Upgrading terraform to 0.10.2 and cloudfoundry provider to 0.9.1
This required to invoke terraform init (as a replacement for terraform get: init also handles fetching modules).
"The downloaded plugins are installed to a subdirectory of the working directory, and are thus local to that working directory" according to https://www.terraform.io/docs/commands/init.html#plugin-installation
The pipeline only commits the tfstate file to git, so it should ignore this directort.

Updated spec files to expect the .terraform/ directory holding plugins and modules and added a SKIP_REMOTE_FLY_DOWNLOAD constant to ease tests debugging