pipeline(depls): rework ops-files and vars-files support
Script manifest-lifecycle-generation handle copy of operators-file. It
is not possible to use the spruce --go-patch option, because we don't
know which template to associate to the ops-file.
In this version, it is not possible to ops-file to directly patch
credentials (use variable as workaround)

Readme extract :
#### ops-files
By convention, all files in template dir matching ```*-operators.yml```
 are used by ```bosh-deployment``` as ```ops-files``` inputs. Theses
 files **are not processed by spruce**.

#### vars-files
By convention, all files in template dir matching ```*-vars-tpl.yml```
are processed by spruce and generate ```*-vars.yml``` files.
Then files are used by ```bosh-deployment``` as ```vars-files``` inputs.

close #33