task(terraform_*): switch to custom image with providers and bump TF 0.11.2
We no longer used official image provided by hashicorp. We use a custom
image with all providers already installed, so we don't need internet
access to download providers. But it seems not possible to forbid download
and allow download for providers not present in the image. So a list of
providers is pre-loaded, to avoid download only.
TF 0.11.2 also requires approval on `apply`, so we skip it using
`auto-approve`.
The docker image is a customization of AlphaGov: see [1]
We also ensure the terraform-provider-cloudfoundry is the one expected.

[1]: https://github.com/orange-cloudfoundry/paas-docker-cloudfoundry-tools/tree/master/terraform