task(git_reset_wip): includes `--include-ignored` to fix tests after 3.8 upgrade
As we want to upload a git repo in out test we have to disable gitignore exclude
This is due to a change introduce in in fly 3.7
```
The execute command will now default to -x, which has been replaced with a new flag, --include-ignored, to revert to the old behavior.

In addition, Fly will no longer blow up when trying to execute with an input that doesn't have a .gitignore. It will also not blow up if any inputs are files and not directories.
```

(cherry picked from commit 7eade5b)