task(generate_depls): restore failure on error
adding a pipe to generate-depls.rb execution was a bad idea, it caused errors to be ignored. So we implement an other mechanism : double tee redirect. One for stdin and another for stderr. To avoid errors being only lost in concourse log, we display error log file at the end of the task.