feature(multi-concourse-version-compliant-pipelines): update pipelines to be able to run it on concourse 5.8.x and concourse 6.4.x
It seems concourse 6.4 do some checks on resource get params and it fails if unexpected params are detected.
We need to support two concourse-pipeline resources as authentication between concourse 5.8.1 and 6.4.x are not compatible.
So we add dual mode to reach concourse, as concourse 6 is the target version, we first try to deploy on v6 and only on failure we try on v5. As when there is a failure, build stop we need to use a `try` step to allow failure, and we need a way to detect success as `try` step always succeed. We use `on_success`, as it seems only triggered when `try` step succeed.
As COA CI, is still using concourse 5.8.1, we need to update IT accordingly.
We need to adapt related tests as `try` step 'hides' previously defined task ('concourse-for-<root-deployment>'