pipeline(bosh-precompile): fix pipeline team
The default team in this pipeline is '<root-deployment-name>' unless it is 'main' in other pipelines. This is required as this pipeline in not loaded as others pipelines, it loaded by upgrade pipeline, not describe in `ci-deployment-overview.yml`.
We include compiled boshreleases in init step.
We use teams in ITs and  auto detected teams declared in reference data set to init credhub.